add move building content to archive
