+++ automation/__init__.py
@@ -1,4 +1,3 @@
-from ntpath import join
 import automation.newsletter as newsletter
 import automation.issues as issues
-import automation.environment as engine
+import automation.engine as engine
